requirement.txt
torch
torchvision
matplotlib
tqdm
google-cloud-pubsub
aiokafka
json 
PIL